\name{amode}
\title{Heterogenous Linear instrumental variables using BART}
\alias{amode}

\description{ Returns an \eqn{\alpha}{alpha} concentration
  parameter corresponding to the number of clusters mode
  being set to \code{Imode}. }

\usage{
amode(
    n, Imode
)
}

\arguments{
   \item{n}{ The number of observations. }
   \item{Imode}{ The number of clusters mode. }
}

\author{
Robert McCulloch: \email{robert.e.mcculloch@gmail.com},\cr
Rodney Sparapani: \email{rsparapa@mcw.edu}.
}
